% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{journalclub.annotate}
\alias{journalclub.annotate}
\title{Retrieve information for a selection of publications}
\usage{
journalclub.annotate(pmids, retmax = 100)
}
\value{
A data frame with one row per publication.
}
\description{
Generates a data frame providing information (abstract, title, authors, journal, publication date)
for a list of PMIDs.
}
