% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{journalclub.candidates}
\alias{journalclub.candidates}
\title{Retrieve new publications for a journal club}
\usage{
journalclub.candidates(dir, query = NULL, recent = 30, debug = F)
}
\arguments{
\item{dir}{Directory where the journal club history is stored.}

\item{query}{PubMed query for identifying new publications (Default: NULL).}

\item{recent}{Number of days to look back for publications from the current day (Default: 30).}
}
\value{
A data frame listing recent publications
that have not been considered for previous journal clubs
and either match the input query or cite a previously presented publication.
}
\description{
Retrieve new publications for a journal club
}
