% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{journalclub.citing}
\alias{journalclub.citing}
\title{Query PubMed for recent citing publications}
\usage{
journalclub.citing(pmids, days = 30, retmax = 100, verbose = T)
}
\arguments{
\item{pmids}{PMIDs for papers to be cited.}

\item{days}{Number of days to look back for publications from the current day (Default: 30).}
}
\value{
A vector of PMIDs matching the query.
}
\description{
Retrieve recent publications that cite a given set of publications
}
