% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{journalclub.query}
\alias{journalclub.query}
\title{Query PubMed for recent publications}
\usage{
journalclub.query(query, days = 30, retmax = 1e+05)
}
\arguments{
\item{query}{PubMed query.}

\item{days}{Number of days to look back for publications from the current day (Default: 30).}
}
\value{
A vector of PMIDs matching the query.
}
\description{
Retrieve PMIDs for recent publications matching a query.
}
