% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{journalclub.update}
\alias{journalclub.update}
\title{Update journal club history with newly presented papers}
\usage{
journalclub.update(dir, presented)
}
\arguments{
\item{dir}{Directory where the journal club history is stored.}

\item{presented}{List of PMIDs for the papers that were presented.}
}
\value{
TRUE
}
\description{
Update journal club history with newly presented papers
}
