% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{retrieve.citing.papers}
\alias{retrieve.citing.papers}
\title{retrieve pmids for papers that cite
the papers with input pmids within the last
few (default: 30) days}
\usage{
retrieve.citing.papers(pmids, days = 30, retmax = 100, verbose = T)
}
\description{
retrieve pmids for papers that cite
the papers with input pmids within the last
few (default: 30) days
}
