% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{retrieve.journal.club.candidates}
\alias{retrieve.journal.club.candidates}
\title{Retrieve new papers for a journal club}
\usage{
retrieve.journal.club.candidates(dir, query = NULL, recent = 30)
}
\arguments{
\item{dir}{Directory where the journal club history is stored.}

\item{query}{PubMed query for identifying new papers (Default: NULL).}

\item{recent}{Number of days to look back for publications from the current day (Default: 30).}
}
\value{
A data frame listing recently published papers
that have not been considered for previous journal clubs
and either match the input query or cite a previously presented paper.
}
\description{
Retrieve new papers for a journal club
}
