% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{retrieve.papers}
\alias{retrieve.papers}
\title{Retrieve information (title, authors, abstract, etc)
for the given paper pmids.
`retmax` limits the number of papers to be retrieved
at a time (but does not limit the total number of papers
to retrieve).}
\usage{
retrieve.papers(pmids, retmax = 100)
}
\description{
Retrieve information (title, authors, abstract, etc)
for the given paper pmids.
`retmax` limits the number of papers to be retrieved
at a time (but does not limit the total number of papers
to retrieve).
}
