% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{retrieve.pmids.by.query}
\alias{retrieve.pmids.by.query}
\title{retrieve pmids for papers
that match the input query
within the few (default: 30) days}
\usage{
retrieve.pmids.by.query(query, days = 30, retmax = 1e+05)
}
\description{
retrieve pmids for papers
that match the input query
within the few (default: 30) days
}
