% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalclub.r
\name{update.journal.club.history}
\alias{update.journal.club.history}
\title{Update journal club history with newly presented papers}
\usage{
\method{update}{journal.club.history}(dir, presented)
}
\arguments{
\item{dir}{Directory where the journal club history is stored.}

\item{presented}{List of PMIDs for the papers that were presented.}
}
\description{
Update journal club history with newly presented papers
}
